+++ linkedin-clone/src/api.js
@@ -1,4 +1,4 @@
-export const apiKey = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NWIyMmVlYTkxM2Y2NTAwMThkMDk1ZDgiLCJpYXQiOjE3MDYxNzYyMzQsImV4cCI6MTcwNzM4NTgzNH0.DCIC1UzDrf0zvwOOcPWGVO2wskfHhWA45iUZ2ClkQB8'
+export const apiKey = 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NWIyMmVlYTkxM2Y2NTAwMThkMDk1ZDgiLCJpYXQiOjE3MDYxNzYyMzQsImV4cCI6MTcwNzM4NTgzNH0.DCIC1UzDrf0zvwOOcPWGVO2wskfHhWA45iUZ2ClkQB8'
 export const endpointProfile = 'https://striveschool-api.herokuapp.com/api/profile/'
 export const endpointPosts = 'https://striveschool-api.herokuapp.com/api/posts'
 export const endpointJobs = 'https://strive-benchmark.herokuapp.com/api/jobs'
